fix the edit event listener and refactor code
