+++ lib/index.js
@@ -1,10 +1,7 @@
 require('./styles.scss')
-require('./event_listeners/addFood.js')
-require('./event_listeners/editFood.js')
-require('./event_listeners/deleteFood.js')
-require('./event_listeners/listAllFoods.js')
+require('./event_listeners/foodEvents.js')
 const $ = require('jquery')
-const meal_events = require('./event_listeners/meal-events')
+const meal_events = require('./event_listeners/mealEvents')
 $(document).ready(() => {
   meal_events;
 });
